requirements.txt
flask==3.0.2
python-dotenv==1.0.1
multidict==6.0.5
pydantic==2.6.4
requests==2.31.0
supabase==2.4.0
yacron==0.19.0
zulip==0.9.0
pytest==8.1.1